+++ packages/orq-rc/src/orq_ai_sdk/models/getbudgetop.py
@@ -161,5 +161,5 @@ class GetBudgetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-11-05T15:06:45.230Z")
+    updated: Optional[datetime] = parse_datetime("2025-11-06T05:48:57.514Z")
     r"""The date and time the resource was last updated"""