+++ packages/orq-rc/src/orq_ai_sdk/models/listcontactsop.py
@@ -186,7 +186,7 @@ class Data(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-11-05T15:06:44.874Z")
+    updated: Optional[datetime] = parse_datetime("2025-11-06T05:48:57.124Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")