+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K9A8X0TZS4S1KTF2D7N9K29R"
+        "01K9BVCC5619PQ8906P0SQ8FEM"
     )
     r"""The unique identifier of the data source"""